Remove let_chains feature
